refactor: update import references to use consolidated prompt system
- Change imports from mantis.prompts to mantis.prompt in executor.py
- Update narrator.py to use consolidated prompt system
- Fix team/base.py imports to use proper A2AExecutor location
- All prompt composition functionality now imports from single module
